requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
jinja2
python-dotenv
google-genai
Pillow
